style (refactored _jove_rt_signal_handler)

+++ lib/jove.macros.h
@@ -42,7 +42,7 @@
         "JOVE UNREACHABLE: \"" __VA_ARGS__ "\" "                               \
         "(" _STRINGIZE(__FILE__) ":" _STRINGIZE(__LINE__) "\n";                \
                                                                                \
-    _jove_sys_write(2 /* stderr */, &__msg[0], sizeof(__msg));                 \
+    _jove_sys_write(2 /* stderr */, (unsigned long)&__msg[0], sizeof(__msg));  \
                                                                                \
     _jove_sys_exit_group(1);                                                   \
     __builtin_unreachable();                                                   \
